refactor: waitForStorage uses resend() instead of getStreamLast(), method moved to Resends.ts
