+++ packages/client/src/subscribe/Resends.ts
@@ -2,9 +2,9 @@
  * Public Resends API
  */
 import { DependencyContainer, inject, Lifecycle, scoped, delay } from 'tsyringe'
-import { MessageRef, StreamPartID, StreamPartIDUtils, StreamID, EthereumAddress } from 'streamr-client-protocol'
+import { MessageRef, StreamPartID, StreamPartIDUtils, StreamID, EthereumAddress, StreamMessage } from 'streamr-client-protocol'
 
-import { instanceId, counterId } from '../utils'
+import { instanceId, counterId, wait } from '../utils'
 import { Context, ContextError } from '../utils/Context'
 import { inspect } from '../utils/log'
 
@@ -18,6 +18,7 @@ import { StreamIDBuilder } from '../StreamIDBuilder'
 import { StreamDefinition } from '../types'
 import { StreamEndpointsCached } from '../StreamEndpointsCached'
 import { range } from 'lodash'
+import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
 
 const MIN_SEQUENCE_NUMBER_VALUE = 0
 
@@ -79,7 +80,8 @@ export default class Resend implements Context {
         @inject(StreamIDBuilder) private streamIdBuilder: StreamIDBuilder,
         @inject(delay(() => StreamEndpointsCached)) private streamEndpoints: StreamEndpointsCached,
         @inject(Rest) private rest: Rest,
-        @inject(BrubeckContainer) private container: DependencyContainer
+        @inject(BrubeckContainer) private container: DependencyContainer,
+        @inject(ConfigInjectionToken.Root) private config: StrictStreamrClientConfig
     ) {
         this.id = instanceId(this)
         this.debug = context.debug.extend(this.id)
@@ -247,5 +249,67 @@ export default class Resend implements Context {
             publisherId,
             msgChainId,
         })
+    }
+
+    async waitForStorage(streamMessage: StreamMessage, {
+        // eslint-disable-next-line no-underscore-dangle
+        interval = this.config._timeouts.storageNode.retryInterval,
+        // eslint-disable-next-line no-underscore-dangle
+        timeout = this.config._timeouts.storageNode.timeout,
+        count = 100,
+        messageMatchFn = (msgTarget: StreamMessage, msgGot: StreamMessage) => {
+            return msgTarget.signature === msgGot.signature
+        }
+    }: {
+        interval?: number
+        timeout?: number
+        count?: number
+        messageMatchFn?: (msgTarget: StreamMessage, msgGot: StreamMessage) => boolean
+    } = {}) {
+        if (!streamMessage) {
+            throw new ContextError(this, 'waitForStorage requires a StreamMessage, got:', streamMessage)
+        }
+
+        const [ streamId, partition ] = StreamPartIDUtils.getStreamIDAndPartition(streamMessage.getStreamPartID())
+        const streamDefinition = { streamId, partition}
+
+        /* eslint-disable no-await-in-loop */
+        const start = Date.now()
+        let last: StreamMessage[]
+        // eslint-disable-next-line no-constant-condition
+        let found = false
+        while (!found) {
+            const duration = Date.now() - start
+            if (duration > timeout) {
+                this.debug('waitForStorage timeout %o', {
+                    timeout,
+                    duration
+                }, {
+                    streamMessage,
+                    last: last!.map((l: any) => l.content),
+                })
+                const err: any = new Error(`timed out after ${duration}ms waiting for message: ${inspect(streamMessage)}`)
+                err.streamMessage = streamMessage
+                throw err
+            }
+
+            const resendStream = await this.resend(streamDefinition, { last: count })
+            last = await resendStream.collect()
+            for (const lastMsg of last) {
+                if (messageMatchFn(streamMessage, lastMsg)) {
+                    found = true
+                    this.debug('last message found')
+                    return
+                }
+            }
+
+            this.debug('message not found, retrying... %o', {
+                msg: streamMessage.getParsedContent(),
+                'last 3': last.slice(-3).map(({ content }: any) => content)
+            })
+
+            await wait(interval)
+        }
+        /* eslint-enable no-await-in-loop */
     }
 }
